# static/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: Mechanic Workshop API
  description: API documentation for Customers, Mechanics, Service Tickets, and Inventory.
# API server configuration specifying host, base path, and HTTP scheme
host: localhost:5000
basePath: /api
schemes:
  - http
# Security definition for JWT Bearer token authentication via Authorization header
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: >
      JWT Token authentication. Use 'Bearer <your-token>' format.
# Tags categorizing API endpoints by resource type
...
paths:
...
  /customers:
    post:
      tags:
        - Customers
      summary: Create Customer
      description: Registers a new customer account
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CustomerInput'
      responses:
        201:
          description: Customer created successfully
          schema:
            $ref: '#/definitions/Customer'
          examples:
            application/json:
              id: 1
              name: "Clark Kent"
              email: "clark@example.com"
              address: "123 Main St, Metropolis, NY"
              phone: "+1-555-123-4567"
# Retrieve a paginated list of all registered customers
  /customers/:
    get:
      tags:
        - Customers
      summary: Find All Customers (Paginated)
      description: Returns a paginated list of customers.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Results per page (default: 10)"
      responses:
        200:
          description: A paginated list of customers.
          schema:
            type: object
            properties:
              customers:
                type: array
                items:
                  $ref: '#/definitions/Customer'
              total:
                type: integer
                example: 50
              pages:
                type: integer
                example: 5
              current_page:
                type: integer
                example: 1
# Authenticate a customer and return a JWT token
  /customers/login:
    post:
      tags:
        - Customers
      summary: Customer Login (Token Auth)
      description: "Logs in a customer and returns a JWT token along with customer details."
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
            example:
              email: "clark@example.com"
              password: "supersecret123"
      responses:
        200:
          description: "Login successful"
          schema:
            type: object
            properties:
              access_token:
                type: string
              token_type:
                type: string
              customer:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
            examples:
              application/json:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type: "Bearer"
                customer:
                  id: 1
                  name: "Clark Kent"
# Retrieve a customer's details by their ID (requires authentication)
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Find Customer by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Customer data
          schema:
            $ref: '#/definitions/Customer'
# Update a customer's information by their ID (requires authentication)
    put:
      tags:
        - Customers
      summary: Update Customer by ID
      description: "Updates an existing customer account."
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: ID of the customer to update
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CustomerInput'
            example:
              name: "Clark Kent"
              email: "clark@example.com"
              password: "newpassword123"
              address: "456 Elm St, Metropolis, NY"
              phone: "+1-555-987-6543"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: '#/definitions/Customer'
          examples:
            application/json:
              id: 1
              name: "Clark Kent"
              email: "clark@example.com"
              address: "456 Elm St, Metropolis, NY"
              phone: "+1-555-987-6543"
# Delete the logged-in customer's account by ID
    delete:
      tags: [Customers]
      summary: Delete Customer by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: ID of the customer to delete
      responses:
        204:
          description: Customer deleted successfully
        401:
          description: Unauthorized – invalid or missing token
        404:
          description: Customer not found
# Retrieve all service tickets associated with the authenticated customer
  /customers/my-tickets:
    get:
      tags: [Customers]
      summary: Find Service Tickets for Authenticated Customer
      security:
        - BearerAuth: []
      responses:
        200:
          description: A list of service tickets
          schema:
            type: array
            items:
              $ref: '#/definitions/ServiceTicket'
# Create a new mechanic and return the created mechanic details
...
definitions:
  Customer:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      address:
        type: string
      phone:
        type: string
    required:
      - name
      - email
      - address
      - phone
# Input model for creating or updating a customer with required fields
  CustomerInput:
    type: object
    required: [name, email, password, address, phone]
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      address:
        type: string
      phone:
        type: string
# Mechanic object model representing a mechanic’s profile and details
...
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
      description:
        type: string
      customer_id:
        type: integer
      vin:
        type: string
    required:
      - id
      - description
      - customer_id
      - vin   
# Input model for creating or updating a service ticket with required fields
...